claude/enhanced/server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
Flask-Limiter==3.5.0
bcrypt==4.0.1
requests==2.31.0
python-dotenv==1.0.0
